backend/requirements.txt
Flask
Flask-Cors
pymongo
langchain
openai
python-dotenv
